LayoutTests/media/wireless-playback-media-player/timeupdate-without-position-updates.html

<!DOCTYPE html><!-- webkit-test-runner [ WirelessPlaybackMediaPlayerEnabled=true ] -->
<html>
    <head>
        <script src='../media-file.js'></script>
        <script src='../video-test.js'></script>
        <script>
            var route;
            var video;

            async function start()
            {
                if (window.internals)
                    internals.mockMediaDeviceRouteController.enabled = true;

                route = internals.mockMediaDeviceRouteController.createMockMediaDeviceRoute();
                run(`internals.mockMediaDeviceRouteController.activateRoute(route)`);

                const urlPromise = new Promise((resolve) => {
                    route.setURLCallback(async (url) => {
                        resolve(url);
                    });
                });

                video = document.getElementsByTagName('video')[0];
                run(`video.src = findMediaFile('video', '../content/test')`);
                run(`route.playbackRate = 1`);
                run(`route.timeRange = { start: 0, duration: 60 }`);
                run(`video.play()`);
                await urlPromise;

                const canplaythroughPromise = waitFor(video, 'canplaythrough');
                run(`route.ready = true`);
                await canplaythroughPromise;

                await waitFor(video, 'timeupdate');
                testExpected('video.currentTime > 0', true);

                endTest();
            }
        </script>
    </head>
    <body onload='start()'>
        <video controls></video>
        <p>Test that the CMTimebase drives timeupdate and advances currentTime even without route position updates.</p>
    </body>
</html>
