LayoutTests/media/wireless-playback-media-player/timeupdate-without-position-updates-expected.txt


Test that the CMTimebase drives timeupdate and advances currentTime even without route position updates.

RUN(internals.mockMediaDeviceRouteController.activateRoute(route))
RUN(video.src = findMediaFile('video', '../content/test'))
RUN(route.playbackRate = 1)
RUN(route.timeRange = { start: 0, duration: 60 })
RUN(video.play())
RUN(route.ready = true)
EVENT(canplaythrough)
EVENT(timeupdate)
EXPECTED (video.currentTime > 0 == 'true') OK
END OF TEST
